a_list_size returns a size_t; various related changes.

+++ src/a2x_pack_console.c
@@ -29,7 +29,7 @@ typedef struct ALine {
 bool g_enabled;
 bool g_show;
 static AList* g_lines;
-static int g_linesPerScreen;
+static size_t g_linesPerScreen;
 static ASprite* g_titles[A_CONSOLE_MAX];
 static AInput* g_toggle;
 
@@ -128,7 +128,7 @@ void a_console__init(void)
     g_enabled = true;
     g_show = false;
     g_lines = a_list_new();
-    g_linesPerScreen = INT_MAX;
+    g_linesPerScreen = SIZE_MAX;
 }
 
 void a_console__init2(void)
