+++ src/a2x_pack_input.p.h
@@ -30,7 +30,7 @@ extern AInput* a_input_new(const char* Names);
 extern const char* a_input_name(const AInput* Input);
 extern bool a_input_working(const AInput* Input);
 
-extern int a_input_numControllers(void);
+extern size_t a_input_numControllers(void);
 extern void a_input_setController(size_t Index);
 
 extern bool a_button_get(const AInput* Button);
